add more tests and add refactor the code accordingly

+++ spec/test_helper.rb
@@ -1,7 +1,7 @@
 # frozen-string-literal: true
 
 require 'minitest/autorun'
-require 'himawari'
+require_relative '../lib/himawari.rb'
 
 class HimawariTest < Minitest::Test
   # rubocop:disable Style/ClassVars
@@ -10,15 +10,33 @@ class HimawariTest < Minitest::Test
 
   def setup
     @himawari = Himawari::Download
+    date1 = Time.parse('2019-01-01 00:00:00 +00:00')
+    date2 = Time.parse('2020-06-01 00:00:00 +00:00')
+    @timestamp = Time.at((date2.to_f - date1.to_f) * rand + date1.to_f)
+
+    # 1200secs == 20.minutes 02:40 and 14:40 are "bad" timestamps because
+    # we know 100% that himawari does not take photos at those times
+    @timestamp -= 1200 if @timestamp.min == 40 && [2, 14].include?(@timestamp.hour)
   end
 
   # def teardown
   #  puts 'run after each test'
   # end
+
+  def capture_stdout
+    original_stdout = $stdout  # capture previous value of $stdout
+    $stdout = StringIO.new     # assign a string buffer to $stdout
+    yield                      # perform the body of the user code
+    $stdout.string             # return the contents of the string buffer
+  ensure
+    $stdout = original_stdout  # restore $stdout to its previous value
+  end
+
   def self.cleanup
-    puts "After Test Cleanup. Will remove #{@@workdir}"
-    `rm -r #{@@workdir}/data*`
+    puts "Cleanup: Removing #{@@workdir}/data"
+    `rm -r #{@@workdir}/data*` if File.directory?("#{@@workdir}/data")
   end
 end
 
+HimawariTest.cleanup
 Minitest.after_run { HimawariTest.cleanup }
